feat: add new binding type enum value 'versionTag'
The new binding type will work similar to 'latest' as a version tag will be resolved to the latest deployed version
that uses this tag. We decided to use a dedicated type though to make the difference more explicit and keep things
similar to what users might already know from Camunda 7.